+++ lunes_cms/api/v2/serializers/job_serializer.py
@@ -15,7 +15,7 @@ class JobSerializer(serializers.ModelSerializer):
 
     @extend_schema_field(OpenApiTypes.BOOL)
     def get_migrated(self, obj):
-        """Return True if the job was migrated from the old data model """
+        """Return True if the job was migrated from the old data model"""
         return obj.v1_id is not None
 
     class Meta: